+++ docs/conf.py
@@ -134,6 +134,7 @@ autoapi_root = "api"
 autoapi_add_toctree_entry = False
 autoapi_ignore = [
     "*/**/_version.py",
+    "*/generated/*",
 ]
 autoapi_options = [
     "members",